+++ src/ui/first_run/default_paths.rs
@@ -12,7 +12,6 @@ pub struct DefaultPathsApp {
     progress_bar: AsyncController<ProgressBar>,
 
     show_additional: bool,
-    migrate_installation: bool,
     show_progress: bool,
 
     launcher: PathBuf,
@@ -49,8 +48,7 @@ pub enum DefaultPathsAppMsg {
 
 #[relm4::component(async, pub)]
 impl SimpleAsyncComponent for DefaultPathsApp {
-    /// If `true`, then use migrate installation mode
-    type Init = bool;
+    type Init = ();
     type Input = DefaultPathsAppMsg;
     type Output = FirstRunAppMsg;
 
@@ -242,11 +240,7 @@ impl SimpleAsyncComponent for DefaultPathsApp {
                     set_spacing: 8,
 
                     gtk::Button {
-                        set_label: &if model.migrate_installation {
-                            tr!("migrate")
-                        } else {
-                            tr!("continue")
-                        },
+                        set_label: &tr!("continue"),
 
                         set_css_classes: &["suggested-action", "pill"],
 
@@ -254,16 +248,9 @@ impl SimpleAsyncComponent for DefaultPathsApp {
                     },
 
                     gtk::Button {
-                        set_label: &if model.migrate_installation {
-                            tr!("close", { "form" = "noun" })
-                        } else {
-                            tr!("exit")
-                        },
+                        set_label: &tr!("exit"),
 
                         add_css_class: "pill",
-
-                        #[watch]
-                        set_visible: !model.migrate_installation,
 
                         connect_clicked => DefaultPathsAppMsg::Exit
                     }
@@ -287,7 +274,7 @@ impl SimpleAsyncComponent for DefaultPathsApp {
         }
     }
 
-    async fn init(init: Self::Init, root: Self::Root, _sender: AsyncComponentSender<Self>) -> AsyncComponentParts<Self> {
+    async fn init(_init: Self::Init, root: Self::Root, _sender: AsyncComponentSender<Self>) -> AsyncComponentParts<Self> {
         let model = Self {
             progress_bar: ProgressBar::builder()
                 .launch(ProgressBarInit {
@@ -299,7 +286,6 @@ impl SimpleAsyncComponent for DefaultPathsApp {
                 .detach(),
 
             show_additional: false,
-            migrate_installation: init,
             show_progress: false,
 
             launcher: LAUNCHER_FOLDER.to_path_buf(),
@@ -311,12 +297,13 @@ impl SimpleAsyncComponent for DefaultPathsApp {
             fps_unlocker: CONFIG.game.enhancements.fps_unlocker.path.clone(),
             components: CONFIG.components.path.clone(),
 
-            #[allow(clippy::or_fun_call)]
-            temp: CONFIG.launcher.temp.clone().unwrap_or(std::env::temp_dir())
+            temp: CONFIG.launcher.temp.clone()
+                .unwrap_or_else(std::env::temp_dir)
         };
 
         // Set progress bar width
-        model.progress_bar.widget().set_width_request(400);
+        model.progress_bar.widget()
+            .set_width_request(400);
 
         let widgets = view_output!();
 
@@ -364,49 +351,9 @@ impl SimpleAsyncComponent for DefaultPathsApp {
 
             #[allow(unused_must_use)]
             DefaultPathsAppMsg::Continue => {
-                let old_config = Config::get().unwrap_or_else(|_| CONFIG.clone());
-
                 match self.update_config() {
                     Ok(_) => {
-                        if self.migrate_installation {
-                            self.progress_bar.sender().send(ProgressBarMsg::SetVisible(true));
-
-                            self.show_progress = true;
-
-                            let folders = [
-                                (old_config.game.wine.builds, &self.runners),
-                                (old_config.game.dxvk.builds, &self.dxvks),
-                                (old_config.game.wine.prefix, &self.prefix),
-                                (old_config.game.path.global, &self.game_global),
-                                (old_config.game.path.china,  &self.game_china),
-                                (old_config.components.path,  &self.components),
-
-                                (old_config.game.enhancements.fps_unlocker.path, &self.fps_unlocker)
-                            ];
-
-                            #[allow(clippy::expect_fun_call)]
-                            for (i, (from, to)) in folders.iter().enumerate() {
-                                self.progress_bar.sender().send(ProgressBarMsg::UpdateCaption(Some(
-                                    from.to_str().map(|str| str.to_string()).unwrap_or_else(|| format!("{:?}", from))
-                                )));
-
-                                if &from != to && from.exists() {
-                                    move_files::move_files(from, to).expect(&format!("Failed to move folder: {:?} -> {:?}", from, to));
-                                }
-
-                                self.progress_bar.sender().send(ProgressBarMsg::UpdateProgress(i as u64 + 1, folders.len() as u64));
-                            }
-
-                            // Restart the app
-
-                            std::process::Command::new(std::env::current_exe().unwrap()).spawn().unwrap();
-
-                            relm4::main_application().quit();
-                        }
-
-                        else {
-                            sender.output(Self::Output::ScrollToSelectVoiceovers);
-                        }
+                        sender.output(Self::Output::ScrollToSelectVoiceovers);
                     }
 
                     Err(err) => {
@@ -419,14 +366,7 @@ impl SimpleAsyncComponent for DefaultPathsApp {
             }
 
             DefaultPathsAppMsg::Exit => {
-                if self.migrate_installation {
-                    // TODO: this shit should return message to general preferences component somehow to close MigrateInstallation window
-                    todo!();
-                }
-
-                else {
-                    relm4::main_application().quit();
-                }
+                relm4::main_application().quit();
             }
         }
     }